requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
# Editable install with no version control (clip==1.0)
-e /home/tudooh/qty/CLIP-main
colorlog==6.7.0
comm==0.1.3
conda-pack==0.6.0
ConfigArgParse==1.5.3
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cryptography @ file:///croot/cryptography_1677533068310/work
cupy-cuda11x==12.0.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
# Editable install with no version control (dassl==0.4.2)
-e /home/c2pcil/xw/PointCLIP_V2-main/zeroshot_cls/Dassl3D
debugpy==1.6.7
decorator==5.1.1
docker-pycreds==0.4.0
easydict==1.9
embreex==2.17.7.post3
entrypoints==0.3
executing==1.2.0
fastjsonschema==2.16.3
fastrlock==0.8.1
filelock==3.12.2
flake8==3.7.9
Flask==2.2.3
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.0.1
future==0.18.3
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1671623667463/work
gdown==4.7.1
gensim==4.3.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.17.3
google-auth-oauthlib==1.0.0
googledrivedownloader==0.4
grpcio==1.54.0
h5py==3.6.0
huggingface-hub==0.16.4
hydra-core==0.11.3
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.4.1
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
iopath==0.1.9
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
isodate==0.6.1
isort==4.3.21
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
kmeans-pytorch==0.3
lightly==1.4.1
lightly-utils==0.0.2
lightning-utilities==0.8.0
lmdb==1.3.0
lxml==4.9.3
mapbox-earcut==1.0.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib @ file:///croot/matplotlib-suite_1679593461707/work
matplotlib-inline==0.1.6
mccabe==0.6.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
multidict==6.0.4
munkres==1.1.4
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==3.1
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
nvidia-ml-py==11.525.112
nvitop==1.1.2
oauthlib==3.2.2
omegaconf==1.4.1
open3d==0.16.0
packaging==23.1
pandas==2.0.0
parso==0.8.3
path @ file:///opt/conda/conda-bld/path_1641578212155/work
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
plotly==5.14.1
ply==3.11
plyfile==0.9
# Editable install with no version control (pointnet==0.0.1)
-e /home/c2pcil/qty/pointnet.pytorch-master
pointnet2 @ file:///home/c2pcil/xw/ULIP-main/Pointnet2_PyTorch-master.zip
portalocker @ file:///tmp/build/80754af9/portalocker_1617133067286/work
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycollada==0.7.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyflakes==2.1.1
Pygments==2.15.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-louvain==0.16
pytorch-lightning==0.7.1
# Editable install with no version control (pytorch-loss==0.0.0)
-e /home/tudooh/qty/pytorch-loss-master
pytorch-ranger==0.1.1
# Editable install with no version control (pytorch3d==0.7.3)
-e /home/tudooh/qty/pytorch3d-main
pytz==2023.3
PyYAML @ file:///croot/pyyaml_1670514731622/work
pyyaml_env_tag==0.1
pyzmq==25.0.2
rdflib==7.0.0
regex==2022.10.31
requests @ file:///croot/requests_1678709721434/work
requests-oauthlib==1.3.1
rsa==4.9
Rtree==1.0.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.34.0
setproctitle==1.3.3
shapely==2.0.2
shortuuid==1.0.11
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.2
svg.path==6.3
tabulate @ file:///opt/conda/conda-bld/tabulate_1657784105888/work
tb-nightly==2.16.0a20231110
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.4.12
tokenizers==0.13.3
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.12.1
torch-geometric==1.7.2
torch-optimizer==0.3.0
torch-scatter==2.1.2
torchaudio==0.12.1
# Editable install with no version control (torchmetrics==0.12.0.dev0)
-e /home/tudooh/qty/torchmetrics-master
torchpq==0.3.0.5
torchvision==0.13.1
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets==5.9.0
transformers==4.31.0
trimesh==4.0.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1680254681959/work
utils==1.0.1
wandb==0.13.3
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.7
xxhash==3.4.1
yacs @ file:///tmp/build/80754af9/yacs_1634047592950/work
yapf==0.29.0
yarl==1.8.2
zipp==3.15.0
